unity/WebApplication/requirements.txt
django
djangorestframework
requests
django-cors-headers
pillow